+++ ocr/include/huaweicloud/ocr/v1/model/LicensePlateResult.h
@@ -49,7 +49,7 @@ public:
     void setPlateNumber(const std::string& value);
 
     /// <summary>
-    /// 当前版本支持的车牌底色类型：  - blue: 蓝色  - green: 绿色  - black: 黑色  - white: 白色  - yellow: 黄色 
+    /// 当前版本支持的车牌底色类型：  - blue: 蓝色  - green: 绿色（小型新能源车牌）  - black: 黑色  - white: 白色  - yellow: 黄色  - yellow_green: 黄绿（大型新能源车牌） 
     /// </summary>
 
     std::string getPlateColor() const;